(refactor) separate mongo models from connection
