+++ server/database/models/index.js
@@ -1,0 +1,5 @@
+require('../');
+
+module.exports.Movie = require('./Movie');
+module.exports.User = require('./User');
+module.exports.Review = require('./Review');
